\name{row263}
\docType{data}
\alias{row263}
\title{Example signal sequence}
\description{
  Sequence of length 512 corresponding to the 263rd row of \code{\link{trees}}.
  The sequence was used as an example in Evers
  and Heaton (2009).
}
\usage{row263}
\examples{
data(row263)
plot(row263, type="b")
}
\seealso{\code{\link{trees}}}
\references{
Evers, L. and Heaton, T. (2009) Locally Adaptive Tree-Based Thresholding. Journal of Computational and Graphical Statistics.  Dec 2009, Vol. 18, No. 4: 961-977.
}
\keyword{datasets}
