\name{row263}
\docType{data}
\alias{row263}
\title{Example signal sequence}
\description{
  Sequence of length 512 corresponding to the 263rd row of \code{\link{trees}}.
  The sequence was used as an example in Evers
  and Heaton (2009).
}
\usage{row263}
\examples{
data(row263)
plot(row263, type="b")
}
\seealso{\code{\link{trees}}}
\references{
Evers, L. and Heaton T. (2009) Locally adaptive tree-based tresholding.
}
\keyword{datasets}
